Vinisha, Supriya | MOBN-1347 | Refactor CONTEXT to be a global constant
