+++ src/apps/concepts/constants.ts
@@ -5,9 +5,3 @@ export const DICTIONARY_VERSION_CONTAINER = "dictionaryVersion";
 export const SOURCE_CONTAINER = "source";
 
 export const FILTER_SOURCE_IDS = Object.keys(PREFERRED_SOURCES);
-
-export const CONTEXT = {
-  create: "create",
-  view: "view",
-  edit: "edit"
-};
