+++ src/utils/constants.ts
@@ -297,3 +297,9 @@ export const PREFERRED_SOURCES_VIEW_ONLY: { [key: string]: string } = {
   "Public Sources": ALL_PUBLIC_SOURCES_URL,
   ...PREFERRED_SOURCES,
 };
+
+export const CONTEXT = {
+  create: "create",
+  view: "view",
+  edit: "edit"
+};
